Move mvsData/imageAlgo.hpp to image module
