+++ src/aliceVision/image/imageAlgo.cpp
