+++ src/aliceVision/image/imageAlgo.hpp
