+++ src/aliceVision/mvsData/CMakeLists.txt
@@ -3,7 +3,6 @@ set(mvsData_files_headers
   Image.hpp
   geometry.hpp
   geometryTriTri.hpp
-  imageAlgo.hpp
   imageIO.hpp
   jetColorMap.hpp
   Matrix3x3.hpp
@@ -25,7 +24,6 @@ set(mvsData_files_headers
 set(mvsData_files_sources
   jetColorMap.cpp
   Image.cpp
-  imageAlgo.cpp
   imageIO.cpp
   geometry.cpp
   geometryTriTri.cpp
@@ -40,6 +38,7 @@ alicevision_add_library(aliceVision_mvsData
   PUBLIC_LINKS
     aliceVision_system
     aliceVision_numeric
+    aliceVision_image
     ZLIB::ZLIB
     Boost::filesystem
     Boost::boost
